Added files relevant for weather study. work in progress
Includes two data sets, one bus, one weather. In the code should be the info required to generate the csv for use with the ML algorithms. So far a multilayer perceptron neural network is performing the best. Going to try random forests, tune parameters, and probably see about running again with bus route and peak/off peak back in.